% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/history.R
\name{assistant_history}
\alias{assistant_history}
\title{Get conversation history}
\usage{
assistant_history(n = NULL, as_messages = TRUE)
}
\arguments{
\item{n}{Integer. Number of recent messages to return. NULL = all.}

\item{as_messages}{Logical. If TRUE, returns in API message format.}
}
\value{
A list of messages.
}
\description{
Get conversation history
}
