% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{assistant_set_model}
\alias{assistant_set_model}
\title{Set the active model}
\usage{
assistant_set_model(model)
}
\arguments{
\item{model}{Character. Model identifier.}
}
\description{
Set the active model
}
