% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/history.R
\docType{data}
\name{HISTORY_MAX_DEFAULT}
\alias{HISTORY_MAX_DEFAULT}
\title{Conversation History Management}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
HISTORY_MAX_DEFAULT
}
\description{
Store and retrieve conversation history for multi-turn dialogues.
}
\keyword{datasets}
