% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addin.R
\docType{data}
\name{.chat_app_env}
\alias{.chat_app_env}
\title{RStudio Addins for R Assistant}
\format{
An object of class \code{environment} of length 1.
}
\usage{
.chat_app_env
}
\description{
RStudio Addins for R Assistant
}
\keyword{datasets}
